Add files via upload
I had created this simple and basic portfolio website which is for beginners and this is the first version of this project. I had used Boxicons library for using the icons of Facebook, WhatsApp, and GitHub. And this project is designed especially for those who are starting there journey in react. I had only used static design in this project and there is nothing related to backend. I will Upload the second and the final version of this project as soon as possible.